% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_object.R
\name{new_dist}
\alias{new_dist}
\title{Create a distribution object}
\usage{
new_dist(name, params, dfun, pfun, qfun, rfun)
}
\arguments{
\item{name}{Character name of the distribution (e.g. "gamma").}

\item{params}{Named list of parameters for the base R functions.}

\item{dfun, pfun, qfun, rfun}{The base R d/p/q/r functions.}
}
\value{
A list with class "distsfactory_dist" containing d/p/q/r methods
and a \code{$params} element.
}
\description{
Create a distribution object
}
\keyword{internal}
