% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{parse_spec}
\alias{parse_spec}
\title{Parse keyword arguments into a specification object}
\usage{
parse_spec(
  mean = NULL,
  var = NULL,
  std = NULL,
  cv = NULL,
  scv = NULL,
  second_moment = NULL,
  median = NULL,
  q1 = NULL,
  q3 = NULL,
  iqr = NULL,
  quantiles = NULL,
  mode = NULL
)
}
\arguments{
\item{mean}{Target mean.}

\item{var}{Target variance.}

\item{std}{Target standard deviation.}

\item{cv}{Coefficient of variation.}

\item{scv}{Squared coefficient of variation.}

\item{second_moment}{Target second raw moment.}

\item{median}{Target median.}

\item{q1}{First quartile.}

\item{q3}{Third quartile.}

\item{iqr}{Interquartile range.}

\item{quantiles}{List of two \code{c(p, q)} vectors.}

\item{mode}{Target mode.}
}
\value{
A list with class set to the spec type.
}
\description{
Parse keyword arguments into a specification object
}
\keyword{internal}
