% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_object.R
\name{print.distsfactory_dist}
\alias{print.distsfactory_dist}
\title{Print a distsfactory distribution}
\usage{
\method{print}{distsfactory_dist}(x, ...)
}
\arguments{
\item{x}{A \code{distsfactory_dist} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns the \code{distsfactory_dist} object \code{x}.
Called for its side effect of printing a description of the distribution
to the console.
}
\description{
Print a distsfactory distribution
}
